Start refactoring outage plugin.
This plugin now uses measurements to determine outages and should provide much cleaner bounds.

This plugin tracks the last received measurement time for each box. If that time is greater than 1 minute, an outage incident is generated. When the box comes back online, the outage is updated with the timestamp of the first received packet.